+++ docs/CHANGELOG.md
@@ -1,4 +1,6 @@
-# Unreleased
+# [Unreleased]
+
+## Added:
 
 - 添加用于执行定时任务的 CronJobExt